Consolidate types and constants.
